core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@5122 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/persist/TableSpaceManagerBlocks.java
@@ -189,7 +189,7 @@ public class TableSpaceManagerBlocks implements TableSpaceManager {
         }
 
         if (pos < Integer.MAX_VALUE) {
-            lookup.add((int) pos, rowSize);
+            lookup.add(pos, rowSize);
 
             freeBlockSize += rowSize;
         }
@@ -244,7 +244,7 @@ public class TableSpaceManagerBlocks implements TableSpaceManager {
         lookup.remove(index);
 
         if (difference > 0) {
-            int pos = key + ((int) rowSize / scale);
+            long pos = key + (rowSize / scale);
 
             lookup.add(pos, difference);
         }
